rename controller file
