+++ src/controllers/fetchPushSubscriptions/index.ts
@@ -1,4 +1,4 @@
 import FetchPushSubscription from "./fetchPushSubscription";
-import CreateFetchPushSubscriptionController from "./createFetchPushSubscription";
+import CreateFetchPushSubscriptionController from "./createFetchPushSubscriptionController";
 
 export { FetchPushSubscription, CreateFetchPushSubscriptionController };
